requirements.txt
Django~=3.2.19
psycopg2-binary==2.9.5
python-dotenv==1.0.0
whitenoise==6.4.0
redis~=4.5.5
sorl-thumbnail>=3.3.0
pillow>=8.3.2
django-braces>=1.15.0
django-taggit>=1.5.0

django-ckeditor~=6.3.2
